==> 01-Presentation.Rmd <==
---
output:
  pdf_document: default
  html_document: default
---
\mainmatter

# Presentation

##	Background and Objectives
Soil nutrient availability can affect ecosystem carbon cycling, plant phenology, plant diversity and community composition, plant-herbivore and plant-soil-microbe interactions, as well as the structure of trophic food webs [@vanSundert2020]. Thus, the broad range of effects of nutrient availability also affects ecosystem functioning in face of global changes, for instance the response of plants to elevated levels of CO~2~ [@vicca2018]. 

In the context of agriculture, the availability of nutrients regulates crop productivity and, as a result, influences food production. Nevertheless, ongoing conflicts and the rise in extreme weather events have led to historically high input prices. This trend is now endangering the progress and achievement of Sustainable Development Goal (SDG) 2, which strives to eliminate hunger by 2030. Soil nutrient data, which serves as a cornerstone for informed choices, now faces a critical juncture, essential for steering agricultural practices towards a sustainable future.

To date, around 2.3 billion people are affected by moderate and severe food insecurity [@FAO2022]. Despite soil nutrient status and availability being vital to the provisioning of ecosystem services, globally-accessible and harmonised datasets on soil nutrient stocks and soil properties that govern nutrient availability are missing. 

Therefore, the current global situation requires an increase of food production while preserving natural (soil) resources, lowering greenhouse gas emissions and optimising the use of agricultural inputs such as fertilisers [@eisenstein2020]. Fertiliser prices more than doubled within one year and grain prices increased by around 25 percent (Jan. 2021 - Jan. 2022) [@ifpri2020]. With the start of the armed conflict in Ukraine in February 2022, this trend became more pronounced.
Growing food insecurity and rapidly increasing fertiliser prices underscore the urgent need for informed decision-making and optimised soil nutrient management. However, a large data gap exists in regards to soil nutrient stocks and soil properties that govern nutrient availability. Therefore, FAO’s Global Soil Partnership (GSP) has launched the Global Soil Nutrient and Nutrient Budget map (GSNmap) initiative in an endeavour to provide harmonised soil nutrient data and information to stakeholders following a country-driven (bottom-up) approach.
Up-to-date soil data on the status and spatial trends of soil nutrients and related soil attributes is key to guide decision makers to close yield gaps, optimize soil nutrient management and protect local natural resources. Therefore, locally-specific information on the status of soil nutrients are needed [@cunningham2013]. The soil information collected in the GSNmap thereby serves as a cornerstone in delineating priority areas for action and thereby seizes the opportunity to reduce food insecurity, close yield gaps, and reduce environmental costs arising from mismanagement of soil nutrients.

## Global Soil Partnership
The Global Soil Partnership (GSP) was established in December 2012 as a mechanism to develop a strong interactive partnership and to enhance collaboration and generate synergies between all stakeholders to raise awareness and protect the world’s soil resources. From land users to policymakers, one of the main objectives of GSP is to improve governance and promote sustainable management of soils. Since its creation, GSP has become an important partnership platform where global soil issues are discussed and addressed by multiple stakeholders at different levels. 

The mandate of GSP is to improve governance of the planet’s limited soil resources in order to guarantee productive agricultural soils for a food-secure world. In addition, it supports other essential soil ecosystem services in accordance with the sovereign right of each Member State over its natural resources. In order to achieve its mandate, GSP addresses six thematic action areas  to be implemented in collaboration with its regional soil partnerships (Figure 1).

The area of work on Soil Information and Data (SID) of the GSP builds an enduring and authoritative global system (GloSIS) to monitor and forecast the condition of the Earth's soil resources and produce map products at the global level. The International Network of Soil Information Institutions (INSII) consists of nationally mandated institutions and GSP partners working together to develop GloSIS. These institutions possess the necessary technical expertise to create and exchange specific national soil information and data. Whether they are sub-national or national, regional or global, all institutions have the opportunity to join the INSII network and contribute to the collection and dissemination on the status of soil resources.

## Country-driven approach and tasks

The GSNmap initiative is jointly implemented by the International Network of Soil Information Institutions (INSII) and the GSP Secretariat. The process is country-driven, involving and supporting all GSP members in developing their national GSNmap data products. The GSNmap products are being developed following a two-phase approach:

* Phase I: development of soil nutrient and associated soil property maps;
* Phase II: quantification, analysis, projections of nutrient budgets for croplands. 

This Technical Manual focuses on the specific methodology to implement Phase I of the GSNmap. It covers the generation of soil property maps for the soil attributes specified in table \@ref(tab:overviewprop). The methodology presented in this document was reviewed and endorsed by INSII and is part of the Country guidelines and technical specifications for global soil nutrient and nutrient budget maps [@FAO2022b].



```{r overviewprop, echo = FALSE, message=F, warning=F}

library(readxl)
library(kableExtra)
dt <- read_xlsx("tables/wosis_dist.xlsx")
dt <- dt[!(dt$`Soil property` %in%c( "P Bray I","P Olsen" )),]
dt[dt$`Soil property` =='P Mehlich 3',"Soil property"] <- "Available P"

kbl(booktabs = T, dt[,c("Soil property","property_id","Unit" )], col.names = gsub("_", " ", names(dt[,c("Soil property","property_id","Unit" )])), caption = 'Mandatory soil attributes and units of the phase I Global Soil Nutrient and Nutrient Budget Maps') %>%
kable_classic(full_width = F)
```
Depending on national data availability and technical capacities, ad-hoc solutions will be developed by the GSNmap WG to support countries during the national GSNmap production and/or harmonisation phase. Where possible, GSP Secretariat will use publicly available data to gap-fill the areas which are not covered by the national submissions unless the country requests to be left blank on the GSNmap products.

## How to use this book

The present document is a technical manual on the phase I of the GSNmap initiative. It provides the scientific background on the importance of soil nutrients and guidance on the digital soil mapping techniques to map nutrients and soil properties that govern nutrient availability. It also comprises a compendium with all necessary scripts to generate national GSNmaps. These scripts are described step-by-step in 4 steps that cover soil data preparation (Step 1), covariate download (Step 2), the mapping process itself (Step 3), and the automatic generation of national reports (Step 4). The general workflow is shown in Figure \@ref(fig:steps).

```{r steps, echo = FALSE, fig.cap = "Overview of the steps to follow for the GSNmap generation.",out.width='100%'}
knitr::include_graphics("images/Manual-Workflow.png")

```

The chapters are structured as following:

* Chapter 1 provides general information about the GSNmap initiative as well as FAO's Global Soil Partnership (GSP).
* Chapter 2 focuses on the scientific state-of-the-art in terms of soil nutrients and soil nutrient mapping.
* Chapter 3 and 4 introduce the software requirements and the concept of digital soil mapping.
* Chapter 5 to 7 guide the reader through the nutrient mapping exercise of GSNmap Phase I providing step-by-step instructions.
* Chapter 8 explains how the national GSNmaps are reported to the GSP.
* Annex I serves as a repository for the complete scripts needed for the GSNmap.
* Annex II provides alternative step-by-step instructions for the special case of soil data without point coordinates.

The GSNmap Technical Manual is structured as a practical document to be used by national experts in the endeavour to employ digital soil mapping techniques to generate national nutrient maps based on a common methodology. The concept of digital soil mapping presented here can however be also used in mapping exercises that focus on other soil properties and is therefore also relevant to scientists and digital soil mappers. The training material and the folders of the technical manual can be downloaded as .zip file here: https://github.com/FAO-GSP/GSNmap-TM/archive/refs/heads/main.zip. Alternatively, the GitHub repository can be cloned to your local device by using the following link: https://github.com/FAO-GSP/GSNmap-TM.git.

## Training material

The train material of this book is located in the [GSNmap-TM GitHub repository](https://github.com/FAO-GSP/GSNmap-TM). To download the input files and **R** scripts, clone the repository or click on [this link](https://github.com/FAO-GSP/GSNmap-TM/archive/refs/heads/main.zip), save the ZIP file and extract its content in a folder, preferable close to the root of your system, such as ```"C:/GIT/"```.
