==> 10-Way-Forward.Rmd <==
# Way forward

This technical manual provided step-by-step guidance on how to generate nutrient maps by means of quantile regression forest models within a digital soil mapping framework. The array of maps produced belongs to the first implementation phase of the GSNmap initiative and provides urgently needed data on nutrient stocks and soil properties that govern nutrient availability. Policymakers will be able to use this data to derive conclusions on where to concentrate efforts to improve soil and land management to strengthen agrifood systems.
The second phase of the GSNmap will make use of the first phase data products to derive nutrient budget maps. Therefore, additional datasets will be used for calculating input and output terms of nutrient stocks. The methodology is currently under development by the GSNmap working group.

## Frequent asked questions and Troubleshooting answers

To be developed soon...

## Issues in the GSNmap Technical Manual

Please, report any issue in the GSNmap Technical Manual in its issues GitHub page [https://github.com/FAO-GSP/GSNmap-TM/issues](https://github.com/FAO-GSP/GSNmap-TM/issues).

## Get help

- Check the issues GitHub page [https://github.com/FAO-GSP/GSNmap-TM/issues](https://github.com/FAO-GSP/GSNmap-TM/issues)
- Issues with R packages: search for solutions in [https://stackoverflow.com/](https://stackoverflow.com/)
- ```caret``` package [https://topepo.github.io/caret/](https://topepo.github.io/caret/)
- ```terra``` package [https://rspatial.org/terra/pkg/1-introduction.html](https://rspatial.org/terra/pkg/1-introduction.html)
- ```tidyverse``` package [https://r4ds.had.co.nz/](https://r4ds.had.co.nz/)
- ```sf``` package [https://r-spatial.github.io/sf/](https://r-spatial.github.io/sf/)
